[feature-make-query-database-universal] Moved the function PrepareThenExecuteQuery() to a newly created package database_util.
